skid code..

+++ commands/stop.js
@@ -1,7 +1,7 @@
 const { SlashCommandBuilder } = require('discord.js');
 const { useQueue } = require('discord-player');
 
-console.log("Loaded play.js")
+console.log("Loaded stop.js")
 module.exports = {
     data: new SlashCommandBuilder()
         .setName('stop')
@@ -13,8 +13,13 @@ module.exports = {
         const channel = member.voice.channel;
 
         if (!channel) {
-            await interaction.reply("えー流したくないなぁー...だってVCに実行者が居ないんだもん...")
+            await interaction.reply("えー実行したくないなぁー...だってVCに君が居ないんだもん...")
+			return;
         }
+
+		const queuedTracks = queue.tracks.toArray();
+    	if (!queuedTracks[0])
+    	  return interaction.reply({ content: `再生されている曲がありません！`, ephemeral: true });
 
         queue.delete();
         await interaction.reply(`音楽を停止しました👋`)
